refact: Clarify 'belong_to_segment' algorithm
